like&hashTag

+++ jango/Day11/instargram/article/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render,redirect
 from django.http.response import HttpResponse
-from .models import Article ,Comment, AticlesImages, Board
+from .models import Article ,Comment, AticlesImages, Board ,HashTag
 import json
 # Create your views here.
 
@@ -51,40 +51,61 @@ def edit_boards(request):
 
 def index(request):
     if request.method=="POST":
-        article = Article()
-        article.contents = request.POST["contents"]
-        #원본이미지를 받을떄
-        #article.image = request.FILES["image"]
-        #원본 이미지를 프로세싱 한 이미지를 저장 
-        #article.image_resized = request.FILES["image"]
-        article.save()
-        for image in request.FILES.getlist("image"):
-            AticlesImages.objects.create(article_id=article.id, image=image)
-        return redirect('articles')
+        if request.user.is_authenticated:
+            article = Article()
+            article.contents = request.POST["contents"]
+            article.user_id = request.user.id
+            article.save()
+            tags = request.POST["hashtags"]
+            for tag in tags.split(","):
+                tag = tag.strip()
+                if not HashTag.objects.filter(tag=tag):
+                    tag = HashTag.objects.create(tag=tag)
+                else:
+                    tag = HashTag.objects.filter(tag=tag)[0]
+                article.tags.add(tag)
+
+            #원본이미지를 받을떄
+            #article.image = request.FILES["image"]
+            #원본 이미지를 프로세싱 한 이미지를 저장 
+            #article.image_resized = request.FILES["image"]
+            
+            for image in request.FILES.getlist("image"):
+                AticlesImages.objects.create(article_id=article.id, image=image)
+            return redirect('articles')
+        else:
+            return redirect('accounts:login')
+        
     else:
         articles = Article.objects.all().order_by("created_at").reverse()
         context={
             'articles':articles
-        }
+            }
         return render(request,'index.html',context)
 
 
 def edit(request, article_id):
     article = Article.objects.get(id=article_id)
-    if request.method=='POST':
-        article.contents =request.POST["contents"]
-        article.save()
-        return redirect('articles')
+    if article.is_permitted(request.user_id):
+        if request.method=='POST':
+            article.contents =request.POST["contents"]
+            article.save()
+            return redirect('articles')
+        else:
+            context={
+            'article':article
+            }
+            return render(request,'article/edit.html',context)
     else:
-        context={
-        'article':article
-        }
-        return render(request,'article/edit.html',context)
+        return redirect('articles')
 
 def delete(request, article_id):
     article = Article.objects.get(id=article_id)
-    article.delete()
-    return redirect('articles')
+    if article.is_permitted(request.user.id):
+        article.delete()
+        return redirect('articles')
+    else:
+        return redirect('articles')
 
 
 
@@ -93,36 +114,51 @@ def delete(request, article_id):
 
 def comments(request):
     if request.method =='POST':
-        contents = request.POST["contents"]
-        article_id = request.POST["article_id"]
+        if request.user.is_authenticated :
+            contents = request.POST["contents"]
+            article_id = request.POST["article_id"]
 
-        if request.POST["form_Method"]=="create":
-            comment = Comment()
-        elif request.POST["form_Method"]=="edit":
-            comment_id = request.POST["comment_id"]
-            comment = Comment.objects.get(id=comment_id)
-        comment.contents=contents
-        comment.article_id = article_id
-        comment.save()
-        context={
-            'Method': request.POST["form_Method"],
-            'comment': comment.contents,
-            'comment_id':comment.id,
-            'article_id': comment.article_id,
-        }
-    return render(request,'comment.html',context)
-
+            if request.POST["form_Method"]=="create":
+                comment = Comment()
+                comment.user_id = request.user.id
+            elif request.POST["form_Method"]=="edit":
+                comment_id = request.POST["comment_id"]
+                comment = Comment.objects.get(id=comment_id)
+            if comment.user_id != request.user.id:
+                return HttpResponse('',status=401)
+            else:
+                comment.contents=contents
+                comment.article_id = article_id
+                comment.user_id = request.user.id
+                comment.save()
+                context={
+                    'Method': request.POST["form_Method"],
+                    'comment': comment.contents,
+                    'username': comment.user.username,
+                    'comment_id':comment.id,
+                    'article_id': comment.article_id,
+                }
+                return render(request,'comment.html',context)
+        else:
+            context = {
+                    'status' :401,
+                    'message':'로그인이 필요합니다.'
+            }
+            return HttpResponse(json.dumps(context), status=401, content_type="application/json")
 
 
 def delete_comment(request):
     if request.method == 'POST':
         id = request.POST["comment_id"]
         comment = Comment.objects.get(id=id)
-        comment.delete()
-        context = {
+        if comment.user_id == request.user.id:
+            comment.delete()
+            context = {
             'comment_id' : id
-        }
-        return HttpResponse(json.dumps(context),content_type="application/json")
+            }   
+            return HttpResponse(json.dumps(context),content_type="application/json")
+        else:
+            return HttpResponse('',status=401)
 
 def edit_comment(request, comment_id):
     comment = Comment.objects.get(id=comment_id)
@@ -135,3 +171,19 @@ def edit_comment(request, comment_id):
             'comment':comment
         }
         return render(request, 'comment/edit.html',context)
+
+def likes(request):
+    if request.user.is_authenticated and request.method == "POST":
+        article_id = request.POST["article_id"]
+        article = Article.objects.get(id=article_id)
+        if request.user in article.user_likes.all():
+            article.user_likes.remove(request.user) #좋아요 취소 
+        else:
+            article.user_likes.add(request.user) #좋아요
+        likes_count = len(article.user_likes.all())
+        context={
+        'count': likes_count
+        }
+        return HttpResponse(json.dumps(context),content_type="application/json")
+    else:
+        return HttpResponse('',status=403)
